Move App state to RecipeContext
